cart-service/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlmodel
psycopg2-binary
python-jose[cryptography]
PyJWT
requests
prometheus_client